Fix spikedet.
The detector should now give mostly correct results.

To do:
* review the filter
* multithreading and optimizations
* more tests and benchmark

+++ download-libraries.sh
@@ -4,8 +4,15 @@ git clone --depth 1 https://github.com/clMathLibraries/clFFT.git
 
 git clone --depth 1 https://github.com/google/googletest.git unit-test/googletest
 
-curl http://www.alglib.net/translator/re/alglib-3.10.0.cpp.gpl.zip > alglib.zip
-unzip alglib.zip -dalglib
-rm alglib.zip
-mv alglib/cpp/* alglib
-rmdir alglib/cpp
+if [ -d alglib ];
+then
+   echo "alglib already exists"
+else
+    curl http://www.alglib.net/translator/re/alglib-3.10.0.cpp.gpl.zip > alglib.zip
+	unzip alglib.zip -dalglib
+	rm alglib.zip
+	mv alglib/cpp/* alglib
+	rmdir alglib/cpp
+fi
+
+git clone --depth 1 https://github.com/RLovelett/eigen.git
